use spray-json to return Blog Entries as JSON, use environment variables for configuration with defaults
