+++ build.sbt
@@ -12,6 +12,7 @@ val akkaVersion = "2.3.6"
 libraryDependencies ++= Seq(
   "com.typesafe.akka" %% "akka-actor" % akkaVersion
   ,"io.spray" %% "spray-can" % sprayVersion
+  ,"io.spray" %% "spray-json" % sprayVersion
   ,"io.spray" %% "spray-routing" % sprayVersion
   ,"org.json4s" %% "json4s-native" % "3.2.11"
   // -- Slick --
