+++ src/main/resources/application.conf
@@ -5,7 +5,6 @@ akka {
 
 # check the reference.conf in spray-can/src/main/resources for all defined settings
 spray.can.server {
-  # uncomment the next line for making this an HTTPS example
 #  ssl-encryption = on
   idle-timeout = 30 s
   request-timeout = 10 s
@@ -21,30 +20,37 @@ blog {
   # Application confs:
   #------------------
   app {
-    hostname = "0.0.0.0"
+    hostname = "localhost"
+    hostname = ${?BLOG_APP_HOSTNAME}
     port = 9999
+    port = ${?BLOG_APP_PORT}
   }
 
   # Database configurations:
   #------------------------
   db {
-    url = "jdbc:mysql://localhost:3306/ozoliblog"
-    driver = com.mysql.jdbc.Driver
-
-    user = "scala"
+    user = "dbuser"
+    user = ${?BLOG_DB_USER}
     port = 3306
+    user = ${?BLOG_DB_PORT}
     password = "password"
-    name = "ozoliblog"
+    password = ${?BLOG_DB_PASSWORD}
+    name = "databaseName"
+    name = ${?BLOG_DB_NAME}
     pool {
       maxObjects = 10
       maxIdle = 4
       maxQueueSize = 10
     }
+
+    url = "jdbc:mysql://"${blog.db.host}":"${blog.db.port}"/"${blog.db.name}""
+    driver = com.mysql.jdbc.Driver
   }
 
   # RSS Feed URL configuration:
   #------------------------
   rss {
-    uri = "http://feeds.feedburner.com/OliverCarrsBlog?fmt=xml"
+    uri = "http://feeds.feedburner.com/YourBlog?fmt=xml"
+    uri = ${?BLOG_RSS_URI}
   }
 }
