+++ src/main/scala/io/ozoli/blog/Main.scala
@@ -10,7 +10,7 @@ object Main extends App with DatabaseConfiguration {
   implicit val system = ActorSystem()
 
   // the handler actor replies to incoming HttpRequests
-  val handler = system.actorOf(Props[DemoService], name = "handler")
+  val handler = system.actorOf(Props[BlogService], name = "handler")
 
-  IO(Http) ! Http.Bind(handler, interface = "localhost", port = 8080)
+  IO(Http) ! Http.Bind(handler, interface = conf.getString("blog.app.hostname"), port = conf.getInt("blog.app.port"))
 }
